pyproject.toml
[project]
name = "postagent"
version = "0.1.0"
description = "Encrypted message broker for AI agents — MQTT pub/sub with end-to-end NaCl encryption"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Geoff Huntley" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
    "pynacl>=1.5",
    "paho-mqtt>=2.1",
    "typer>=0.15",
    "httpx>=0.28",
    "base58>=2.1",
    "pydantic>=2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "mypy>=1.14",
]

[project.urls]
Homepage = "https://github.com/geoff616/postagent"
Repository = "https://github.com/geoff616/postagent"

[project.scripts]
postagent = "postagent.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
